Desktop/Nutri_Hospital/requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
psycopg2-binary
reportlab
Flask-Migrate
alembic
pandas
openpyxl
